C++: refactor Errors class
